Fixes chmln/flatpickr#61
Ensured that non-inline elements are appended to the BODY tag so that we
can absolutely position them relative to their corresponding input with script.

This prevents calendars getting hidden inside of things like in-page modal
windows (submodal / lightbox / etc).

There still could probably be some improvements to position the left of
calendar if bumps against a window boundary, but I'll leave that for
another time.