Updated parseConfig
Updated the parseConfig to take into account the altInput field so that the self.config.altFormat is not overwritten if altInput is true.  Without this, altFormat is overwritten whenever self.config.enableTime is true; if you wanted the time selector, then you couldn't have an alternative human friendly view of the time.