Fixed disabled issue when using altInput
This fixes the following bug.

  - Use the ```altInput``` flag in the configuration.
  - Disable the main input
  - The calendar in that scenario does not disable. It still opens when you click on it.

The reason for this is simple. Your check for disabled input checks if the altInput exists and uses it over the main input for the disabled flag. This would work fine if the altInput would clone the disabled property from the main input, but it does not. Because of this, disabled inputs do not work when using the ```altInput = true``` flag. This fixes that problem.