Basic support for server-side-rendering
This checks if `window` exists before checking `window.navigator` in the default config.
Without this check, Flatpickr cannot be loaded in Node.js (for e.g. server-side-rendering the react component).

This has a side-effect of disabling animation in non-browser contexts :P

I also added a very simple test that just require's the Flatpickr.js source file in node.js. A more comprehensive test would build the dist folder and then require that, but this is probably good enough for now. Note that this test must run outside of Jest, because Jest automatically pollyfills window, which would make it pass even when it shouldn't.

This fixes the root cause of the following issues:
* https://github.com/coderhaoxin/react-flatpickr/issues/9
* https://github.com/carbon-design-system/carbon-components-react/issues/76