Fix `data-toggle` button on mobile
Currently the `data-toggle` button doesn't work on mobile. This is due the `.click()` event triggering on the `mobileInput` doesnt show the mobile date picker.
Triggering a focus event with `.focus()` makes the mobile date picker show up as expected.

This commit also enables passing through the `event` param from the `toggle()` method to the `open()` method.