fix: Focus calendar day on open()/toggle()
Added a call to `focusOnDay()` inside `open()` so that triggering the datepicker to open via keyboard focuses the calendar days when days are visible.  This allows keyboard controls to then navigate around the calendar days rather than going straight to the "hours" field.

Included is a change that causes datepickers that only have a time picker available but are using the `wrap: true` option to focus on the "hours" field whenever the time picker is opened via keyboard interaction on the `data-toggle` element.