Make clearing on backspace configurable
In #936 a feature was introduced to make it so that pressing backspace or
delete would clear the input.

For some inputs though, it's desirable to not allow the user to clear
the input. For example, we might want to be able to set a date on load
and only allow modifying that date.

This commit introduces a configuration flag to disable the clearing
feature called `allowClearing`. The field is defaulted to `true` to
preserve the current default behavior.

While I was in here I also added the missing tests for the original behavior,
and the new configuration flag.

Closes #1867