Fix stale year display in monthSelect plugin (#2278)
Fixes #2277[0].

Originally, in the monthSelect plugin,
clicking on the prev/next year buttons
failed to update the year shown at the top of the calendar
until after the user made an initial selection.
Under the hood, clicking the prev/next year buttons did change the year,
meaning that the selected month/year would not necessarily match
the month/year displayed in the UI.

This occurred because the line that updates the view
was hidden behind an `if` clause that it shouldn't have been.
This commit moves that line to the appropriate place.

[0]: https://github.com/flatpickr/flatpickr/issues/2277