Fix monthSelect's .today/.selected styling
This commit corrects a mismatch in the CSS/styling
of the default date select mode
and the monthSelect plugin.

Specifically, when default flatpickr is newly instantiated:

* `.selected` is applied to nothing, and
* `.today` is applied to the current date.

When flatpickr is newly instantiated with the monthSelect plugin:

* `.selected` is applied to the current date, and
* `.today` is applied to nothing.

(Importantly, the current date is not actually selected by default,
meaning that the UI does not reflect the underlying calendar state.)

To fix this, monthSelect's DOM management had to be expanded:
before, it simply drew the calendar grid once,
and paging through previous/future years had no influence
on which cell was marked `.selected`;
now, it clears and redraws the calendar grid with each page change
so that the `.today` and `.selected` cells come and go as needed.