Add monthSelect's range-mode hover effect
This commit was made in preparation for the next one,
which will introduce support for "range" mode in monthSelect.
It adds visual styling only.

With this commit, after a month has been selected in range mode,
flatpickr will close (which is not desired behavior)--
but re-opening it and hovering over other months
will cause the month cells to be highlighted as appropriate,
as a preview of the tentative range selection.

This was accomplished by exposing flatpickr's `onMouseOver` function,
which--despite its generic name--is specifically for implementing
this range-mode hover highlighting functionality.

=== Auxiliary changes

This commit also replaces a for loop in onMouseOver
with a more functional .forEach().

This change was purely cosmetic.