monthSelct plugin: Enable "range" mode
This commit concludes work started in the previous commit--
namely, adding support for "range" mode to the monthSelect plugin.

Where 743b4d2 adds the visual feedback component
(cells should be highlighted in response to clicking/hovering),
this commit adds the functional component,
saving selections to fp.selectedDates (or resetting it) as appropriate.

Where possible, the logic here was written
to mirror that of the base date picker.
This is not the DRYest possible solution,
and leads to the risk of feature drift
(or just more maintenance overhead)
if relevant features/fixes are added to the base date picker
without making corresponding changes in monthSelect.

A cleaner design would require a massive rewrite of the library
in closer adherence to the open-closed principle.[0]
(This commit includes unit tests to support that possibility!)

[0]: https://en.wikipedia.org/wiki/Open%E2%80%93closed_principle