Fix monthSelect plugin's minDate handling
This commit addresses an edge case in an already-fixed bug (3fa5a69)
in the monthSelect plugin's handling of self.config.minDate bounds.

Specifically, the edge case in question occurs only in "range" mode,
after a start date has been selected
and then input is abandoned before the end date is selected.
In these cases, when flatpickr is reopened,
the "previous month" button always shows,
even when the minDate bound has been reached.
(Clicking it is a no-op, but the button shouldn't show at all.)

The previous fix addressed this with the following line of plugin config
(to understand why, consult the commit message):

    fp.currentMonth = 0;

In the edge case, abandoning input invokes self.clear(),
which resets currentMonth back to the actual current month:

    self.currentMonth = self._initialDate.getMonth();

To get around this, this commit overrides _initialdate.getMonth, too:

    fp._initialDate.getMonth = () => 0;

It's hacky, but so was the original fix.
Luckily, there are unit tests, so it can be safely refactored later.