Fix for picker showing and hiding again (#2304)
* Fix for picker showing and hiding again

In some cases, the mousedown and the mouseup events would not have the same target; this is because the calendar gets drawn under the cursor during the focus event, which steals the mousedown event, ultimately meaning a click is run on the document instead of on self._input. This patch updates both the document handler and the input event handler to use mousedown instead of click, ensuring the calendar doesn't flash on and off the screen. This addresses issue #2286.

* Update index.spec.ts