Revert "Merge pull request #1846 from lightsound/update-ja"
This reverts commit 580ec769478c3d879943a5458fb316806035cbb6, reversing
changes made to 9f850d72bf7eea75218de447e46054dbc5e12494.

The first day of the week in japan is Sunday instead of Monday.
There are two reasons.

The first one is that the Kojien (Japanese Dictionary) says that Sunday is
first day of the week.
The second one is that Labor Standards Act in Japan decides the default
week is Sunday to Saturday.

[昭63.1.1 基発1号](https://www.mhlw.go.jp/web/t_doc?dataId=00tb1899&dataType=1&pageNo=1)

Because of these two reasons, The first day of the week is Sunday.