fix: handling document click after manual update
When manual input is allowed, the user input was not respected, when
a user clicks anywhere in the document body instead of pressing TAB key.
This results in 'click'->'blur' event sequence as opposed to only a
blur event. This results in setting the date always from the calendar
element.

The fix compares the date from the input with the latest selected date;
if it matches then the changes is coming from the calendar control and
updates the date time from calendar control, as before. Otherwise the
change is coming from the input control, and it simply overwrites the
latest selected date. The synchronization step later takes care of
the rest, as before.