Fix parseDate on relative inputs
Suppose the day is given as 0. date.setDate(0) will set the day to the last of the previous month. With ops.forEach being inside the for loop this is done for every run through the loop, so for every loop we go one month back instead of only once.

For example parseDate("0.1.2021", "d.m.Y") gives 31.08.2020 instead of 31.12.2020.