Eliminate memory leak caused by event listener accumulation
Prior to this commit, destroying a FlatpickerInstance
would not remove capturing event listeners registered
by that instance. This caused detached event listeners
to accumulate as FlatpickerInstances were created and
destroyed, which caused memory usage to increase.
Capturing event listeners were not removed because the
call to `removeEventListener()` was made without the
`options` argument that indicated it was a capturing
event listener. If an event listener is registered
as a capturing listener, then it must be removed as
a capturing listener[1].

This commit modifies the event listener removal functions
generated by `bind()` to include the `options` argument
when calling `removeEventListener()`.

[1] https://developer.mozilla.org/en-US/docs/Web/API/EventTarget/removeEventListener